alcotest: fix EEXIST on mkdir
Our CI is particularly "good" at finding race conditions in unit tests today for some reason, latest victim is alcotest:
```
unittest.exe: internal error, uncaught exception:
              Unix.Unix_error(Unix.EEXIST, "mkdir", "//builddir/build/BUILD/...../_build/default/test/fake/_build")
```

This is a TOCTOU race condition between checking that the directory does not exist and creating it. Avoid it by ignoring EEXIST on mkdir, if we get EEXIST then we lost the race condition and someone else has already created the directory, but that is fine: all we wanted is to ensure the directory exists in the first place.

Using fix suggested by @lindig.